[SofaGraphComponent] MOVE BackgroundSetting to SofaBaseVisual;
