+++ SofaKernel/modules/SofaBaseVisual/CMakeLists.txt
@@ -3,6 +3,7 @@ project(SofaBaseVisual)
 
 set(HEADER_FILES
     BaseCamera.h
+    BackgroundSetting.h
     Camera.h
     InteractiveCamera.h
     VisualModelImpl.h
@@ -13,6 +14,7 @@ set(HEADER_FILES
 
 set(SOURCE_FILES
     BaseCamera.cpp
+    BackgroundSetting.cpp
     Camera.cpp
     InteractiveCamera.cpp
     VisualModelImpl.cpp
