+++ modules/SofaGraphComponent/CMakeLists.txt
@@ -14,7 +14,6 @@ list(APPEND HEADER_FILES
     AddFrameButtonSetting.h
     AddRecordedCameraButtonSetting.h
     AttachBodyButtonSetting.h
-    BackgroundSetting.h
     FixPickedParticleButtonSetting.h
     Gravity.h
     InteractingBehaviorModel.h
@@ -34,7 +33,6 @@ list(APPEND SOURCE_FILES
     AddFrameButtonSetting.cpp
     AddRecordedCameraButtonSetting.cpp
     AttachBodyButtonSetting.cpp
-    BackgroundSetting.cpp
     FixPickedParticleButtonSetting.cpp
     Gravity.cpp
     MouseButtonSetting.cpp
